[#105] fixing linting, refactoring tree;
