+++ src/ml_service/.gitignore
@@ -13,3 +13,4 @@ coverage.xml
 /feature_extraction/Clustering/dbscan/cache.pickle
 /feature_extraction/Clustering/dbscan/fact_clustering.log
 /feature_extraction/Clustering/dbscan/cluster_dir/*
+/feature_extraction/Clustering/k_means/cluster_dir/*
