+++ src/ml_service/feature_extraction/Preprocessing/Taimoor_Parser/preprocessing.py
@@ -2,11 +2,10 @@ import re
 
 
 def preprocessing(facts):
-
     """
     Takes a list of strings and removes/replaces words such that they map to a common word. 
     Also remove unnecessary words/phrases
-    
+
     :param facts: list of strings 
     :return: sanitized list of strings
     """
@@ -33,15 +32,15 @@ def preprocessing(facts):
         # replace date value
         facts[i] = re.sub("\d+-\d+-\d+", " date ", facts[i])
         facts[i] = re.sub(
-            "\d+.{0,2}?\s+("+months+")(\s+(\d{4}))?",
+            "\d+.{0,2}?\s+(" + months + ")(\s+(\d{4}))?",
             " date ", facts[i])
         facts[i] = re.sub(
-            "("+months+")\s+(\d{4})",
+            "(" + months + ")\s+(\d{4})",
             " date ", facts[i])
 
         # replace month values
         facts[i] = re.sub(
-            "("+months+")",
+            "(" + months + ")",
             " nom_du_mois ", facts[i])
 
         # remove hyphens
